requirements.txt
fastapi[all]==0.101.1
mysql-connector-python==8.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pymysql==1.1.0
sqlalchemy==4.7.1